+++ raisedragon-external/src/main/kotlin/com/whatever/raisedragon/external/aws/s3/S3AgentImpl.kt
@@ -1,11 +1,12 @@
-package com.whatever.raisedragon.aws.s3
+package com.whatever.raisedragon.external.aws.s3
 
 import com.amazonaws.AmazonServiceException
 import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.model.ObjectMetadata
+import com.whatever.raisedragon.external.config.S3Config
 import org.springframework.context.annotation.Profile
+import org.springframework.core.io.InputStreamSource
 import org.springframework.stereotype.Component
-import org.springframework.web.multipart.MultipartFile
 import java.io.IOException
 import java.util.*
 
@@ -15,13 +16,18 @@ class S3AgentImpl(
     private val amazonS3: AmazonS3,
     private val s3Properties: S3Config.S3Properties
 ) : S3Agent {
-    override fun upload(directory: String, file: MultipartFile): String {
-        val fileName = String.format("%s%s-%s", directory, UUID.randomUUID(), file.originalFilename)
+    override fun upload(
+        fileInputStreamSource: InputStreamSource,
+        directoryName: String,
+        originalFileName: String,
+        contentType: String,
+        size: Long
+    ): String {
+        val fileName = String.format("%s%s-%s", directoryName, UUID.randomUUID(), originalFileName)
         val bucket = s3Properties.bucketName
         runCatching {
-            val objectMetadata = getObjectMetadataFromFile(file)
-            val inputStream = file.inputStream
-            amazonS3.putObject(bucket, fileName, inputStream, objectMetadata)
+            val objectMetadata = makeObjectMetadataFromFile(contentType, size)
+            amazonS3.putObject(bucket, fileName, fileInputStreamSource.inputStream, objectMetadata)
         }.onFailure { exception ->
             when (exception) {
                 is IOException -> throw IllegalStateException("S3 File I/O Error", exception)
@@ -34,10 +40,10 @@ class S3AgentImpl(
         return "${s3Properties.domain}/$fileName"
     }
 
-    private fun getObjectMetadataFromFile(imageFile: MultipartFile): ObjectMetadata {
+    private fun makeObjectMetadataFromFile(contentType: String, size: Long): ObjectMetadata {
         return ObjectMetadata().apply {
-            contentType = imageFile.contentType
-            contentLength = imageFile.size
+            this.contentType = contentType
+            this.contentLength = size
         }
     }
 }
