+++ raisedragon-external/src/main/kotlin/com/whatever/raisedragon/external/oauth/AuthService.kt
@@ -1,4 +1,4 @@
-package com.whatever.raisedragon.domain.auth
+package com.whatever.raisedragon.external.oauth
 
 interface AuthService {
     /**
