+++ raisedragon-external/src/main/kotlin/com/whatever/raisedragon/external/oauth/kakao/KakaoLoginResponse.kt
@@ -1,4 +1,4 @@
-package com.whatever.raisedragon.domain.auth.oauth.kakao
+package com.whatever.raisedragon.external.oauth.kakao
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import java.time.LocalDateTime
